cron job and logout updates

+++ api/controller/authorize.js
@@ -2,6 +2,7 @@ const { AuthToken } = require("../../models/authtoken");
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
 require("dotenv/config");
+const SECRET = process.env.SECRET
 
 module.exports = {
   logout: async (req, res) => {
@@ -9,22 +10,28 @@ module.exports = {
     // console.log("tokennnnnn", token)
     // const decode2 = jwt.verify(JSON.parse(token), SECRET);
     // console.log("decode2", decode2)
-    jwt.verify(JSON.parse(token), SECRET,async function (err, decoded) {
-      if (err) {
-          res.status(403).json({success: false, message: err.message});
-      } else {
-          const userId= decoded.userId
-          const resData = await AuthToken.deleteMany({userId});
-          if (resData) {
-            return res
-              .status(200)
-              .send({ success: true, message: "Logout Successfully."});
-          } else {
-            return res
-              .status(200)
-              .send({ success: false, message: "Session Expired."});
+    if(token===null){
+      return res
+      .status(200)
+      .send({ success: true, message: "Logout Successfully."});
+    }else{
+      jwt.verify(JSON.parse(token), SECRET,async function (err, decoded) {
+        if (err) {
+            res.status(403).json({success: false, message: err.message});
+        } else {
+            const userId= decoded.userId
+            const resData = await AuthToken.deleteMany({userId});
+            if (resData) {
+              return res
+                .status(200)
+                .send({ success: true, message: "Logout Successfully."});
+            } else {
+              return res
+                .status(200)
+                .send({ success: false, message: "Session Expired."});
+            }
           }
-        }
-    })
+      })
+    }
   }
 }
